+++ src/build.task.ts
@@ -1,6 +1,5 @@
 import {Timings} from '@feltcoop/felt/util/timings.js';
 import {printTimings} from '@feltcoop/felt/util/print.js';
-import type esbuild from 'esbuild';
 
 import {type Task} from './task/task.js';
 import {loadConfig, type GroConfig} from './config/config.js';
@@ -8,20 +7,17 @@ import {adapt} from './adapt/adapt.js';
 import {buildSource} from './build/buildSource.js';
 import {Plugins} from './plugin/plugin.js';
 import {cleanFs} from './fs/clean.js';
-
-export interface TaskArgs {
-	clean?: boolean;
-	'no-clean'?: boolean;
-	mapBundleOptions?: esbuild.BuildOptions;
-}
+import {type BuildTaskArgs} from './build.js';
+import {BuildTaskArgsSchema} from './build.schema.js';
 
 export interface TaskEvents {
 	'build.createConfig': (config: GroConfig) => void;
 }
 
-export const task: Task<TaskArgs, TaskEvents> = {
+export const task: Task<BuildTaskArgs, TaskEvents> = {
 	summary: 'build the project',
 	production: true,
+	args: BuildTaskArgsSchema,
 	run: async (ctx): Promise<void> => {
 		const {fs, dev, log, events, args} = ctx;