+++ src/format.schema.ts
@@ -1,0 +1,14 @@
+import type {ArgsSchema} from './task/task.js';
+
+export const FormatTaskArgsSchema: ArgsSchema = {
+	$id: '/schemas/FormatTaskArgs.json',
+	type: 'object',
+	properties: {
+		check: {
+			type: 'boolean',
+			default: false,
+			description: 'exit with a nonzero code if any files are unformatted',
+		},
+	},
+	additionalProperties: false,
+};